Encrypt dentry chunks.
Not entirely happy with how this turned out. But it works.
Each dentry now has an associated secret chunk that all
of its chunks are XORed with.

Tricky part is that a "zero" chunk is now equivalent to
the secret chunk for a given dentry.

As an extra, I made a number of dentry related allocations
optional. Dentry chunk trees and secrets are not allocated
until needed. This should save some memory.