Add dentry rename support.
This was a bit tougher than expected, as locking is a little tricky.
If rename is across multiple directories (like move), I needed
to make sure that locking order is preserved. So if the new destination
directory is a decendent of the old one, it needs to be locked first.
Otherwise, if the new directory is an ancestor of the old one, the old
directory needs to be locked first. If the two directories are not
siblings or cousins, then locking is performed based on pointer values.

The actual rename is composed of the following operations:
- allocate new disk_dentry
- make old disk_dentry last one in its parent
- swap dentries
- mark old disk_dentry as deleted

This entire operation needs to be (mostly) atomic wrt to other
rename attempts of the same dentry, and wrt lookup operations.