Start abstracting chunk database code.
Zunkfs now keeps a list of chunk databases in which
to search for chunks. The databases are passed in
via the command line:

$ zunkfs --chunk-db=<ro|rw>,<db spec>

You can pass in multiple chunk databases, and they will
be processed in the order that they're specified. Earlier
databases are treated as caches. That is, if a chunk is
found in the 3rd database, it will be writte to the 2nd and
1st ones.

With this commit, there are two types of db specs available.
The first is "dir:" which specifies a local directory that
is to contain chunks. The second, "cmd:", replaces ZUNKFS_FETCH.

It is now possible to do things like:

$ zunkfs --chunk-db=ro,dir:/path/to/ro/chunkdb \
         --chunk-db=rw,dir:/path/to/rw/chunkdb \
	 --chunk-db=ro,cmd:/path/to/http-fetch.sh \
	 --chunk-db=ro,cmd:/path/to/ftp-fetch.sh