Add preliminary chunk-map support.
Not everybody wants to duplicate their files in chunks.
So this lets them store chunk-maps which tell zunkfs
that a chunk can be found in file XXX at offset YYY.

Currently, only a simple sqlite3 database is used.
The schema for it is:

CREATE TABLE chunk_map (
		hash CHAR(20) PRIMARY KEY,
		path VARCHAR(1024),
		chunk_nr INTEGER
);

The code is setup to allow other maps too.

The db spec for sqlite chunk maps is "map:sqlite:/path/to/db".