Fix file close() and flush().
Due to the cyclic nature of an open file's chunk cache,
it was possible for chunks to remain in the cache, even
if a flush() or close() was called.

This triggered an assert() in from close(), as the dentry's
chunk_tree was still in use, even tho the dentry's refcount
was 0.

There are two ways of fixing it. One was to reset ccache_index,
the other (implemented) was to traverse the entire cache
each time.