First external tool: zunkfs-list-ddents.
Use zunkfs-list-ddents to get detailed information about disk_dentries
in the current directory. The implementation is a bit hackish, as Linux
prevents one from opening a directory as a file for writing (EISDIR).
So each directory now has a 'super-secret' file (.super_secret_file) which
gives direct access to the raw directory data.