Clean up directory-as-a-file handling (aka super-secret-file.)
Most code used find_dentry_super_secret() might as well fold
that into find_dentry(). After that, there are no callers of
find_dentry_parent() with name==NULL and pparent==NULL. So
remove handling of those cases.

Remove references to 'super-secret-file' and call it
dir-as-file, as that's more indicative of actual usage.

Finally, make sure that nobody can try to create a file
that has the same name as the dir-as-file one.