Fix some fundamental problems with zunkdb.
The zunkdb client got a rewrite, with the hope of making it
easier to understand. Not sure if it was successful, but I
kept it.

zunkdb itself had a major issue with 'forwarding' mode.
The problem manifested itself in an infinite loop with
nodes sending the same chunks to each other over and over
again. In some cases the same node would keep sending a chunk
to ITSELF!!

The cause is that a node does not know its own hash. So it
just chooses a set of nodes that are closer to the chunk's
digest than the sender. But for any given pair of nodes,
there may be 2 connections between them. This is because
a node is both a client and a server. So node 'A' has 'B' as
its client, and node 'B' has node 'A' as its client.

Trying to work around it by reducing the number of connections
between nodes introduced other problems. In one instance,
a chunk would get stored only at one node. Which is not desirable.

To fix this issue, I added a new protocol operation, FORWARD_CHUNK.
In response to a FORWARD_CHUNK request, a node will send the chunk
to N closes nodes via STORE_CHUNK. These nodes may reply with
STORE_NODEs and REQUEST_DONE messages. For a STORE_NODE, a node
may end up sending a chunk to the new node, if the new node is
closer to the chunk's digest than any node that the chunk was already
sent to.

It is expected that end clients will use FORWARD_CHUNK to reduce
the load on their outgoing pipe. This is now the default way
that the zunkfs zdb backend works. I also added an option to
use STORE_CHUNK if the user desires.