Reduce socket usage in zunkdb
By allowing client nodes to be promoted to server nodes.
This means that for any pair of nodes (A,B) there should
be at most just one socket connecting them, instead of 2.

But this is not tottaly safe, as any client that sends a
STORE_NODE :<port> message (no ip address) is assumed to
be a node, and will be promoted.

So by default, this is disabled. Users may enabled it via
command line option.