Fix dir-as-a-file read handling.
Change fcfe01811cae35b1f57ddf042254caa1b53d2728 marked individual
chunks as cached by using cnode->_private. But that assumed that
->_private wasn't used for anything. Which is true if the cnode
is just a file. But for directories that's not true. And when
that directory is treated as a file (by zunkfs-*-ddents)
the assert() in is_cached() at file.c:54:

	Assertion `cnode->_private == ((void *)0xf0f0f0f0) || cnode->_private == ((void *)0)' failed.