Introduce PUSH_CHUNK zunkdb message.
FORWARD_CHUNK solved the bandwidth issue for clients, in that a client
only needs to send a chunk to one node. But nodes behind slow uplinks
still have the problem of sending a chunk to many other nodes.

PUSH_CHUNK works around the issue by providing a "max_distance" along
with the chunk. A node that receives a PUSH_CHUNK knows that it should
send this chunk to nodes that are closer to the chunk than "max_distance".

To make sure that chunks do get spread out nicely, before sending a
PUSH_CHUNK, a node will first send STORE_NODE for nodes that it thinks
are closest to the chunk.

Currently, PUSH_CHUNK is not the default method, as it does not deal with
disappearing nodes. PUSH_CHUNK should end up working similarily to
FORWARD_CHUNK, and wait for a REQUEST_DONE message. But that's future stuff.
It's also quite possible that PUSH_CHUNK will replace STORE_CHUNK &
FORWARD_CHUNK.

To enable this new method, specify --slow-uplink to zunkdb's command line.