Change dirty chunk_node tracking.
Don't just set a bit, move dirty chunks onto a dirty list.
Make the list per-ctree, and make flush_chunk_tree() be
much faster. It no longer needs to scan all chunk nodes
to see which ones need to be flushed. This simplified
the code a bit, but didn't shrink it any :(