Don't send a node back to itself.
When processing requests, zunkdb may need to search for nodes close to
the digest given by the request. This list may include the node that
issued the request. But that's bad, because it sends the node back
to itself via STORE_NODE. As nodes don't really know what their external
address is, they may end up connecting to themselves. So make sure that
the issuing node is skipped when looking for nodes near the digest.