More error handling cleanups.
Use stdbool for bool type and lower-case true/false.

As there were no other users of vsprintf_new, fold it into sprintf_new.

For chunk-db ctors, the error strings should not include a \n, and
should be fairly short.

In chunk-db-zdb, getaddrinfo() returns an error code that is not errno.
Need to look it up via gai_errstring().